desktop-server/requirements.txt
pywin32
Pillow
pystray